make sure gpg-agent is available

+++ bashrc
@@ -66,13 +66,15 @@ fi
 # Ensure gpg-agent starts with --enable-ssh-support
 if [[ $EUID -ne 0 ]]; then
     if [[ ! -f /run/user/$(id -u)/gpg-agent.env ]]; then
-        killall gpg-agent &> /dev/null;
-        eval $( \
-            gpg-agent \
-                --daemon \
-                --enable-ssh-support \
-                > /run/user/$(id -u)/gpg-agent.env
-        )
+        if [[ -f /usr/bin/gpg-agent ]]; then
+            killall gpg-agent &> /dev/null;
+            eval $( \
+                gpg-agent \
+                    --daemon \
+                    --enable-ssh-support \
+                    > /run/user/$(id -u)/gpg-agent.env
+            )
+        fi
     fi
 fi
 if [[ -f /run/user/$(id -u)/gpg-agent.env ]]; then
